backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
alembic
requests
tensorflow-cpu==2.15.0
pillow
pyyaml
numpy
opencv-python-headless
gunicorn
